Add alternative solution for challenge six
In the new solution, now called `search_for_solution`, the edit distance
is used to find the key. This solution is much closer to the one recommended in the Cryptopals description. It's also much more efficient.

The initial solution, which uses challenge problem three's English text scoring to score all the possible decryptions from keysizes 2 to 40, and then use the
highest score, is now named `search_for_solution_two`.